Supports single nacos in windows
We can use bash shell and Powershell to manage the lifecycle of nacos in Windows.

However, bash shell included with Git for Windows is used. That means there some Linux Commands cannot support, so there are some strange problem occurred, such as cannot use wget to download file, use sed to update configuration file successfully but failed to run because it still used the previous configuration file to run, cannot run shutdown.sh to destroy the started nacos-server process.

Of cause, we can use Powershell to solve all of these problem, because windows supports Powershell better. Unfortunately, Powershell is very difficult to use

Summary, this PR is incomplete, and the supported functions are

1. Download nacos-server-2.0.3.tar.gz
2. Unpack nacos-server-2.0.3.tar.gz
3. Update cluster to standalone in startup.cmd file
4. Start nacos server use bash shell to run startup.sh

the unsupported functions are

1. Shutdown the started nacos server instance (failed to use shutdown.sh to destroy the nacos server instance)
2. Cannot started the other nacos server instance because of port is aready in used, although we modified the port successfully